Move module to corresponding package
